included an animated hamburger for the menu toggle
