+++ src/containers/instructor/CohortNav.js
@@ -1,5 +1,5 @@
 import { connect } from "react-redux";
-import MenuToggle from "../../components/MenuToggle";
+import SideNav from "../../components/SideNav";
 import { onClickIcon } from "../../data/actions";
 import { getUserRole } from "../../utilities/utilities";
 
@@ -12,4 +12,4 @@ const mapDispatchToProps = dispatch => ({
 	onClickIcon: (id) => dispatch(onClickIcon(id)),
 })
 
-export default connect(mapStateToProps, mapDispatchToProps)(MenuToggle);
+export default connect(mapStateToProps, mapDispatchToProps)(SideNav);
