+++ src/css/input.scss
@@ -139,6 +139,7 @@ p {
 @import 'modules/login';
 @import 'modules/topicsNav';
 @import 'modules/icons';
+@import 'modules/hamburgers/hamburgers';
 @import 'modules/toggleMenu';
 @import 'modules/topic';
 @import 'modules/taskList';
