+++ src/css/modules/toggleMenu.scss
@@ -2,19 +2,17 @@
 	display: none;
 }
 .navNarrow {
-	.narrowLink {
-		display: none;
-		.topics-icon-title {
-			display: block;
-		}
-	}
+
 }
-.center-div {
+.toggle-div {
 	display: flex;
 	align-items: center;
-	justify-content: center;
 }
-.center-icon {
+.toggle-icon {
 	width: 40px;
 	height: 40px;
+}
+.toggle-p {
+	font-size: 1.25em;
+	margin: 0;
 }
